♻️ (#344) TaskDetailStore 타입 파일 분리
