+++ src/features/task-detail/components/CommentSection/CommentItem.tsx
@@ -6,7 +6,7 @@ import type { FileInfo } from '@/features/task-detail/types/taskDetailType';
 import { useAuthStore } from '@/features/auth/store/useAuthStore';
 import { CommentActionsMenu } from '@/features/task-detail/components/CommentSection/CommentActionsMenu';
 import { AuthorAvatar } from './AuthorAvatar';
-import { useTaskDetailStore } from '@/features/task-detail/store/useTaskDetailStore';
+import { useTaskDetailStore } from '@/features/task-detail/store/taskDetail/useTaskDetailStore';
 
 interface CommentItemProps {
   comment: CommentUIType;
