+++ src/features/task-detail/components/PdfViewer/PdfHeaderBar.tsx
@@ -1,9 +1,9 @@
 import { usePdfStore } from '@/features/task-detail/store/usePdfStore';
-import { useTaskDetailStore } from '@/features/task-detail/store/useTaskDetailStore';
+import { useTaskDetailStore } from '@/features/task-detail/store/taskDetail/useTaskDetailStore';
 
 const PdfHeaderBar = () => {
-  const { clearFileState, selectedFile} = useTaskDetailStore();
-const {resetPdf}= usePdfStore()
+  const { clearFileState, selectedFile } = useTaskDetailStore();
+  const { resetPdf } = usePdfStore();
   return (
     <div className="w-full h-12 flex items-center justify-between bg-white border-b border-gray-400 px-4">
       <span className="text-sm font-medium truncate max-w-[70%]">
@@ -12,7 +12,7 @@ const {resetPdf}= usePdfStore()
       <button
         onClick={() => {
           clearFileState();
-          resetPdf()
+          resetPdf();
         }}
         className="text-sm bg-gray-700 hover:bg-gray-800 text-white px-3 py-1 rounded-md transition"
       >
