+++ src/features/task-detail/hooks/usePdfPinInteraction.ts
@@ -1,6 +1,6 @@
 import { useRef } from 'react';
 import type { PinWithAuthor } from '@/features/task-detail/types/taskDetailType';
-import { useTaskDetailStore } from '@/features/task-detail/store/useTaskDetailStore';
+import { useTaskDetailStore } from '@/features/task-detail/store/taskDetail/useTaskDetailStore';
 import type { PageSize } from '@/features/task-detail/types/pdfTypes'; // width, height 타입
 import { useAuthStore } from '@/features/auth/store/useAuthStore';
